Will add logic to WelcomeForm to add current player's name to redux state
