+++ src/containers/WelcomeForm/WelcomeForm.js
@@ -1,19 +1,28 @@
 import React from 'react'
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux'
+import { setPlayer } from '../../actions'
 
 export class WelcomeForm extends React.Component {
     constructor(){
         super()
         this.state = {
-            name
+            name: ''
         };
     };
-
+    
+    
     handleChange = event => {
         this.setState({ [event.target.name]: event.target.value })
     }
-
+    
     handleSubmit = event => {
-
+        event.preventDefault();
+        const { setPlayer } = this.props
+        setPlayer(this.state.name)
+        this.setState({
+            name: ''
+        })
     }
 
     canBeSubmitted() {
@@ -26,13 +35,17 @@ export class WelcomeForm extends React.Component {
         return(
             <div className="welcomeForm__div">
                 <form className="welcomeForm">
-                    <label htmlFor="Name" className="welcomeFormName__label">Name</label>
-                    <input className="welcomeFormName__input" id="Name" type="text" name="Name" placeholder="NAME" onChange={this.handleChange} value={this.state.name} />
-                    <button  type="button" className="welcomeFormName__button" onClick={this.handleSubmit(submit)}>Submit</button>
+                    <label htmlFor="name" className="welcomeFormName__label">Name</label>
+                    <input className="welcomeFormName__input" id="name" type="text" name="name" placeholder="NAME" onChange={this.handleChange} value={this.state.name} />
+                    <button  type="button" className="welcomeFormName__button" onClick={this.handleSubmit}>Submit</button>
                 </form>
             </div>
         )
     }
 };
 
-export default WelcomeForm;
+const mapDispatchToProps = dispatch => (bindActionCreators ({
+    setPlayer
+}, dispatch))
+
+export default connect(null, mapDispatchToProps)(WelcomeForm);
