+++ src/reducers/currentPlayer.js
@@ -1,0 +1,8 @@
+export const currentPlayer = (state = '', action) => {
+    switch(action.type) {
+        case 'SET_PLAYER':
+            return action.playerName
+        default:
+            return state
+    }
+}
