+++ src/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
 import { gameData } from './gameData'
+import { currentPlayer } from './currentPlayer'
 
 export const rootReducer = combineReducers({
-  gameData: gameData
+  gameData,
+  currentPlayer
 })
